[feat] Implement GitHub login
