+++ app/api/auth/[...nextauth]/route.js
@@ -1,0 +1,2 @@
+import { handlers } from "@/app/lib/auth";
+export const { GET, POST } = handlers;
