+++ middleware.js
@@ -1,7 +1,4 @@
-import NextAuth from "next-auth";
-import { authConfig } from "@/app/lib/auth.config";
-
-export default NextAuth(authConfig).auth;
+export { auth as default } from "@/app/lib/auth";
 
 export const config = {
   matcher: ["/((?!api|_next/static|_next/image|.*\\.png$|.*\\.JPG$).*)"],
