+++ package.json
@@ -13,7 +13,6 @@
     "@tailwindcss/aspect-ratio": "^0.4.2",
     "@tailwindcss/forms": "^0.5.7",
     "@vercel/postgres": "^0.9.0",
-    "bcrypt": "^5.1.1",
     "clsx": "^2.0.0",
     "landingai": "^1.0.24",
     "landingai-react": "^1.0.24",
@@ -25,7 +24,6 @@
     "zod": "^3.23.8"
   },
   "devDependencies": {
-    "@types/bcrypt": "^5.0.2",
     "@types/react": "18.3.3",
     "eslint": "^8",
     "eslint-config-next": "14.2.5",
